Fix typo in API request error log
